requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
async-timeout==4.0.3
click==8.1.7
Django==4.2.14
django-graphql-jwt==0.4.0
django-rq==2.10.2
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
graphqlclient==0.2.4
promise==2.3
PyJWT==2.8.0
redis==5.0.7
rq==1.16.2
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
